Remove PaymentCommonModule (#4193)
