+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/injection/FlowControllerComponent.kt
@@ -7,7 +7,7 @@ import androidx.lifecycle.ViewModelStoreOwner
 import com.stripe.android.googlepaylauncher.GooglePayLauncherModule
 import com.stripe.android.payments.core.injection.CoroutineContextModule
 import com.stripe.android.payments.core.injection.InjectorKey
-import com.stripe.android.payments.core.injection.PaymentCommonModule
+import com.stripe.android.payments.core.injection.StripeRepositoryModule
 import com.stripe.android.paymentsheet.PaymentOptionCallback
 import com.stripe.android.paymentsheet.PaymentOptionsViewModel
 import com.stripe.android.paymentsheet.PaymentSheetResultCallback
@@ -21,7 +21,7 @@ import javax.inject.Singleton
 @Singleton
 @Component(
     modules = [
-        PaymentCommonModule::class,
+        StripeRepositoryModule::class,
         PaymentSheetCommonModule::class,
         FlowControllerModule::class,
         GooglePayLauncherModule::class,
