+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/injection/PaymentOptionsViewModelModule.kt
@@ -1,31 +1,18 @@
 package com.stripe.android.paymentsheet.injection
 
 import android.content.Context
-import com.stripe.android.PaymentConfiguration
 import com.stripe.android.payments.core.injection.IOContext
-import com.stripe.android.payments.core.injection.PUBLISHABLE_KEY
 import com.stripe.android.paymentsheet.DefaultPrefsRepository
 import com.stripe.android.paymentsheet.PaymentSheet
 import com.stripe.android.paymentsheet.PrefsRepository
 import com.stripe.android.paymentsheet.analytics.EventReporter
 import dagger.Module
 import dagger.Provides
-import javax.inject.Named
 import javax.inject.Singleton
 import kotlin.coroutines.CoroutineContext
 
 @Module
 internal class PaymentOptionsViewModelModule {
-    @Provides
-    fun providePaymentConfiguration(appContext: Context): PaymentConfiguration {
-        return PaymentConfiguration.getInstance(appContext)
-    }
-
-    @Provides
-    @Named(PUBLISHABLE_KEY)
-    fun providePublishableKey(paymentConfiguration: PaymentConfiguration): () -> String {
-        return { paymentConfiguration.publishableKey }
-    }
 
     @Provides
     @Singleton
