+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/injection/PaymentSheetCommonModule.kt
@@ -1,8 +1,11 @@
 package com.stripe.android.paymentsheet.injection
 
+import android.content.Context
 import androidx.core.os.LocaleListCompat
 import com.stripe.android.Logger
+import com.stripe.android.PaymentConfiguration
 import com.stripe.android.payments.core.injection.ENABLE_LOGGING
+import com.stripe.android.payments.core.injection.PUBLISHABLE_KEY
 import com.stripe.android.paymentsheet.BuildConfig
 import com.stripe.android.paymentsheet.analytics.DefaultDeviceIdRepository
 import com.stripe.android.paymentsheet.analytics.DefaultEventReporter
@@ -12,9 +15,11 @@ import com.stripe.android.paymentsheet.repositories.CustomerApiRepository
 import com.stripe.android.paymentsheet.repositories.CustomerRepository
 import com.stripe.android.paymentsheet.repositories.StripeIntentRepository
 import dagger.Binds
+import dagger.Lazy
 import dagger.Module
 import dagger.Provides
 import javax.inject.Named
+import javax.inject.Provider
 import javax.inject.Singleton
 
 @Module
@@ -35,6 +40,26 @@ internal abstract class PaymentSheetCommonModule {
     ): StripeIntentRepository
 
     companion object {
+        /**
+         * Provides a non-singleton PaymentConfiguration.
+         *
+         * Should be fetched only when it's needed, to allow client to set the publishableKey and
+         * stripeAccountId in PaymentConfiguration any time before configuring the FlowController
+         * or presenting Payment Sheet.
+         *
+         * Should always be injected with [Lazy] or [Provider].
+         */
+        @Provides
+        fun providePaymentConfiguration(appContext: Context): PaymentConfiguration {
+            return PaymentConfiguration.getInstance(appContext)
+        }
+
+        @Provides
+        @Named(PUBLISHABLE_KEY)
+        fun providePublishableKey(paymentConfiguration: PaymentConfiguration): () -> String {
+            return { paymentConfiguration.publishableKey }
+        }
+
         @Provides
         @Singleton
         @Named(ENABLE_LOGGING)
